Added pickaxe_history to help exclude mass moves.
When pinpointing add or removes for asserts, the problem was that we
were seeing commits where there was some big change that caused a bunch
of code to migrate as a block. pygit2's diffs registered that as a bunch
of Removeds and an equal bunch of Addeds, so exluding Assertions with equal
numbers of each for any partiular name(predicate) combo within one file
seems sensible.